+++ core/src/main/java/org/opentele/consult/repository/OrganisationUserRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface OrganisationUserRepository extends AbstractRepository<OrganisationUser> {
     OrganisationUser findByUserAndOrganisation(User user, Organisation currentOrganisation);
-    OrganisationUser findByUserAndOrganisationId(User user, int organisationId);
+    OrganisationUser findByUserAndOrganisationId(User user, long organisationId);
     List<OrganisationUser> findAllByOrganisationName(String organisationName);
     List<OrganisationUser> findAllByOrganisationOrderByUserName(Organisation organisation);
     List<OrganisationUser> findAllByUser(User user);
